+++ documentation/pom.xml
@@ -112,7 +112,7 @@
 									<inline>
 										<files>
 											<file>
-												<source>${project.basedir}/src/main/resources/helloservice.yaml</source>
+												<source>${project.basedir}/src/main/resources/stakit.yaml</source>
 												<destName>${git.commit.id.describe}.yaml</destName>
 											</file>
 											<file>
@@ -155,7 +155,7 @@
 							<goal>generate</goal>
 						</goals>
 						<configuration>
-							<inputSpec>${project.basedir}/src/main/resources/helloservice.yaml</inputSpec>
+							<inputSpec>${project.basedir}/src/main/resources/stakit.yaml</inputSpec>
 							<generatorName>java</generatorName>
 							<configOptions>
 								<sourceFolder>src/test/java</sourceFolder>
@@ -174,7 +174,7 @@
 							<goal>generate</goal>
 						</goals>
 						<configuration>
-							<inputSpec>${project.basedir}/src/main/resources/helloservice.yaml</inputSpec>
+							<inputSpec>${project.basedir}/src/main/resources/stakit.yaml</inputSpec>
 							<generatorName>spring</generatorName>
 							<configOptions>
 								<sourceFolder>src/main/java</sourceFolder>
