--- integrationtest/src/test/java/dk/kvalitetsit/hello/integrationtest/TestApplication.java
@@ -1,11 +1,0 @@
-package dk.kvalitetsit.hello.integrationtest;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-@SpringBootApplication
-public class TestApplication extends SpringBootServletInitializer {
-    public static void main(String[] args) {
-        new ServiceStarter().startServices();
-    }
-}
