+++ integrationtest/src/test/java/dk/kvalitetsit/stakit/integrationtest/TestApplication.java
@@ -1,0 +1,11 @@
+package dk.kvalitetsit.stakit.integrationtest;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+public class TestApplication extends SpringBootServletInitializer {
+    public static void main(String[] args) {
+        new ServiceStarter().startServices();
+    }
+}
