--- service/src/main/java/dk/kvalitetsit/hello/controller/exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package dk.kvalitetsit.hello.controller.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class BadRequestException extends RuntimeException {
-}
