--- service/src/main/java/dk/kvalitetsit/hello/dao/HelloDao.java
@@ -1,11 +1,0 @@
-package dk.kvalitetsit.hello.dao;
-
-import dk.kvalitetsit.hello.dao.entity.HelloEntity;
-
-import java.util.List;
-
-public interface HelloDao {
-    void insert(HelloEntity helloEntity);
-
-    List<HelloEntity> findAll();
-}
