--- service/src/main/java/dk/kvalitetsit/hello/service/model/HelloServiceInput.java
@@ -1,4 +1,0 @@
-package dk.kvalitetsit.hello.service.model;
-
-public record HelloServiceInput(String name) {
-}
