+++ service/src/main/java/dk/kvalitetsit/stakit/configuration/JacksonConfiguration.java
@@ -1,0 +1,14 @@
+package dk.kvalitetsit.stakit.configuration;
+
+import com.fasterxml.jackson.databind.Module;
+import org.openapitools.jackson.nullable.JsonNullableModule;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JacksonConfiguration {
+    @Bean
+    public Module jsonNullableModule() {
+        return new JsonNullableModule();
+    }
+}
