+++ service/src/main/java/dk/kvalitetsit/stakit/service/model/HelloServiceInput.java
@@ -1,0 +1,4 @@
+package dk.kvalitetsit.stakit.service.model;
+
+public record HelloServiceInput(String name) {
+}
