+++ service/src/main/java/dk/kvalitetsit/stakit/service/model/HelloServiceOutput.java
@@ -1,0 +1,6 @@
+package dk.kvalitetsit.stakit.service.model;
+
+import java.time.ZonedDateTime;
+
+public record HelloServiceOutput(String name, ZonedDateTime now) {
+}
