+++ web/src/main/java/dk/kvalitetsit/stakit/Application.java
@@ -1,0 +1,11 @@
+package dk.kvalitetsit.stakit;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+}
