--- web/src/test/java/dk/kvalitetsit/hello/VideoLinkHandlerApplicationTests.java
@@ -1,14 +1,0 @@
-package dk.kvalitetsit.hello;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@Disabled
-class VideoLinkHandlerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-}
