+++ web/src/test/java/dk/kvalitetsit/stakit/ApplicationTests.java
@@ -1,0 +1,14 @@
+package dk.kvalitetsit.stakit;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+@Disabled
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+}
